# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Zubair Traders bakery management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  unit: string;
  purchaseCost: number;
  sellingPrice: number;
  stockQuantity: number;
  minStockAlert: number;
}

export interface TrendPoint {
  label: string;
  value: number;
}

export interface Activity {
  id: number;
  title: string;
  detail: string;
  time: string;
  type: string;
}

export interface Dashboard {
  dailySales: number;
  todaysProfit: number;
  buyerDebt: number;
  supplierOwed: number;
  cashInDrawer: number;
  lowStock: Product[];
  salesTrend: TrendPoint[];
  recentActivity: Activity[];
}

export interface Buyer {
  id: number;
  name: string;
  phone: string;
  cnic: string;
  address: string;
  creditLimit: number;
  currentBalance: number;
  createdAt: string;
}

export interface BuyerInput {
  name: string;
  phone: string;
  cnic: string;
  address: string;
  creditLimit: number;
}

export interface BuyerUpdate {
  name?: string;
  phone?: string;
  cnic?: string;
  address?: string;
  creditLimit?: number;
}

export interface Supplier {
  id: number;
  name: string;
  phone: string;
  companyName: string;
  address: string;
  currentBalance: number;
  createdAt: string;
}

export interface SupplierInput {
  name: string;
  phone: string;
  companyName: string;
  address: string;
}

export interface ProductInput {
  name: string;
  unit: string;
  purchaseCost: number;
  sellingPrice: number;
  stockQuantity: number;
  minStockAlert: number;
}

export interface Sale {
  id: number;
  invoiceNumber: string;
  buyerName: string;
  totalAmount: number;
  paidAmount: number;
  dueAmount: number;
  paymentStatus: string;
  transactionTime: string;
}

export interface SaleInput {
  buyerId: number;
  totalAmount: number;
  paidAmount: number;
  paymentStatus: string;
  notes?: string;
}

export interface Expense {
  id: number;
  category: string;
  amount: number;
  description: string;
  expenseDate: string;
}

export interface ExpenseInput {
  category: string;
  amount: number;
  description: string;
  expenseDate?: string;
}

export interface Loan {
  id: number;
  personName: string;
  phone: string;
  loanType: string;
  amount: number;
  balanceRemaining: number;
  description: string;
  dateGiven: string;
}

export interface LoanInput {
  personName: string;
  phone: string;
  loanType: string;
  amount: number;
  description: string;
}