# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Zubair Traders bakery management API
servers:
  - url: /api
...
paths:
...
  /buyers/{id}:
    patch:
      operationId: updateBuyer
      tags: [buyers]
      summary: Update buyer
      parameters:
        - { name: id, in: path, required: true, schema: { type: number } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BuyerUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Buyer" }
...
components:
  schemas:
...
    Buyer:
      type: object
      required: [id, name, phone, cnic, address, creditLimit, currentBalance, createdAt]
      properties:
        id: { type: number }
        name: { type: string }
        phone: { type: string }
        cnic: { type: string }
        address: { type: string }
        creditLimit: { type: number }
        currentBalance: { type: number }
        createdAt: { type: string }
...
    BuyerUpdate:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        cnic: { type: string }
        address: { type: string }
        creditLimit: { type: number }
...