# lib/api-zod/src/generated/types/buyer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Zubair Traders bakery management API
 * OpenAPI spec version: 0.1.0
 */

export interface Buyer {
  id: number;
  name: string;
  phone: string;
  cnic: string;
  address: string;
  creditLimit: number;
  currentBalance: number;
  createdAt: string;
}